chore: apply tweaks
