+++ packages/ws-server/src/index.ts
@@ -1,1 +1,2 @@
-export * from './ws-rnt-server'
+export * from './data-source'
+export * from './server'
